Add files via upload
This project reviews the current cybersecurity posture of Botium Toys. The audit focused on assets, risks, and existing controls. Findings show that while basic protections (firewalls, antivirus, and physical security) are in place, there are critical weaknesses in data protection, access management, and recovery planning. The overall risk score is high (8/10) due to missing controls and lack of compliance with industry standards